pyproject.toml
[project]
name = "carbon"
version = "1.0.01"
description = ""
authors = [
    {name = "Ayush Kumar",email = "ayush438500@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "babel (>=2.18.0,<3.0.0)",
    "pendulum (>=3.2.0,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "polib (>=1.2.0,<2.0.0)",
    "pyright (>=1.1.408,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.0,<1.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
]
